backend/requirements.txt
sqlalchemy
python-jose[cryptography]
python-dotenv
email-validator
fastapi
uvicorn
python-multipart
mediapipe
numpy
pillow
pydantic
passlib[argon2]
